added base layer add and remove static class
